Fix SwiftLint file length errors
